pyproject.toml
[project]
name = "feedback-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "ruff>=0.9.10",
]
